apps/api/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.17

# Environment and config
python-dotenv==1.0.1

# Validation
pydantic==2.9.2
pydantic-settings==2.5.2

# Video processing dependencies (from batch_core.py)
opencv-python==4.10.0.84
numpy==2.1.3
Pillow==10.4.0
moviepy==2.1.1

# Testing
requests==2.32.3
